+++ test/multisig_payload_test.dart
@@ -1,0 +1,21 @@
+import 'package:test/test.dart';
+import 'package:aptos_api_dart/aptos_api_dart.dart';
+
+// tests for MultisigPayload
+void main() {
+  //final instance = MultisigPayloadBuilder();
+  // TODO add properties to the builder and call build()
+
+  group(MultisigPayload, () {
+    // A hex encoded 32 byte Aptos account address.  This is represented in a string as a 64 character hex string, sometimes shortened by stripping leading 0s, and adding a 0x.  For example, address 0x0000000000000000000000000000000000000000000000000000000000000001 is represented as 0x1.
+    // String multisigAddress
+    test('to test the property `multisigAddress`', () async {
+      // TODO
+    });
+
+    // MultisigTransactionPayload transactionPayload
+    test('to test the property `transactionPayload`', () async {
+      // TODO
+    });
+  });
+}