+++ test/transaction_payload_multisig_payload_all_of_test.dart
@@ -1,0 +1,15 @@
+import 'package:test/test.dart';
+import 'package:aptos_api_dart/aptos_api_dart.dart';
+
+// tests for TransactionPayloadMultisigPayloadAllOf
+void main() {
+  //final instance = TransactionPayloadMultisigPayloadAllOfBuilder();
+  // TODO add properties to the builder and call build()
+
+  group(TransactionPayloadMultisigPayloadAllOf, () {
+    // String type
+    test('to test the property `type`', () async {
+      // TODO
+    });
+  });
+}